KGE_asLM/requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
boto==2.49.0
boto3==1.9.220
botocore==1.12.220
certifi==2019.3.9
chardet==3.0.4
decorator==4.3.2
docutils==0.15.2
gast==0.2.2
gensim==3.8.0
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
numpy==1.16.2
pandas==0.24.2
parso==0.3.4
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
smart-open==1.8.4
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorflow-hub==0.4.0
termcolor==1.1.0
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.0
